+++ packages/rule-schema-to-typescript-types/src/errors.ts
@@ -1,0 +1,17 @@
+export class NotSupportedError extends Error {
+  constructor(thing: string, target: unknown) {
+    super(
+      `Generating a type for ${thing} is not currently supported:\n${JSON.stringify(
+        target,
+        null,
+        2,
+      )}`,
+    );
+  }
+}
+
+export class UnexpectedError extends Error {
+  constructor(error: string, target: unknown) {
+    super(`Unexpected Error: ${error}:\n${JSON.stringify(target, null, 2)}`);
+  }
+}