+++ packages/rule-schema-to-typescript-types/src/isArray.ts
@@ -1,0 +1,4 @@
+// https://github.com/microsoft/TypeScript/issues/17002
+export function isArray(arg: unknown): arg is readonly unknown[] {
+  return Array.isArray(arg);
+}