+++ packages/rule-tester/src/utils/dependencyConstraints.ts
@@ -1,0 +1,48 @@
+import * as semver from 'semver';
+
+import type {
+  DependencyConstraint,
+  SemverVersionConstraint,
+} from '../types/DependencyConstraint';
+
+const BASE_SATISFIES_OPTIONS: semver.RangeOptions = {
+  includePrerelease: true,
+};
+
+function satisfiesDependencyConstraint(
+  packageName: string,
+  constraintIn: DependencyConstraint[string],
+): boolean {
+  const constraint: SemverVersionConstraint =
+    typeof constraintIn === 'string'
+      ? {
+          range: `>=${constraintIn}`,
+        }
+      : constraintIn;
+
+  return semver.satisfies(
+    (require(`${packageName}/package.json`) as { version: string }).version,
+    constraint.range,
+    typeof constraint.options === 'object'
+      ? { ...BASE_SATISFIES_OPTIONS, ...constraint.options }
+      : constraint.options,
+  );
+}
+
+export function satisfiesAllDependencyConstraints(
+  dependencyConstraints: DependencyConstraint | undefined,
+): boolean {
+  if (dependencyConstraints == null) {
+    return true;
+  }
+
+  for (const [packageName, constraint] of Object.entries(
+    dependencyConstraints,
+  )) {
+    if (!satisfiesDependencyConstraint(packageName, constraint)) {
+      return false;
+    }
+  }
+
+  return true;
+}